added a row and column tsx to create a grid for better css control in the transactions table
