+++ app/(tabs)/utilities/balance-utilities.ts
@@ -1,116 +1,20 @@
-// Purpose: Contains utility functions for the balance screen.
 import Toast from "react-native-toast-message";
-import { getFireApp } from "@/getFireApp";
-import { TransactionProps } from "../../types";
+
+
+// Purpose: Contains utility/helper functions for the balance screen.
+
 
 /**
- * Adds a transaction to Firestore.
- * @param amount The amount of the transaction.
- * @param date The date of the transaction.
- * @param description The description of the transaction.
- * @param setAmount A function to set the amount state.
- * @param setDate A function to set the date state.
- * @param setDescription A function to set the description state.
+ * Formats a Firestore Timestamp to a short date string (MM/DD).
+ * @param timestamp The Firestore Timestamp to format.
+ * @returns A formatted date string in MM/DD format.
  */
-export const addTransaction = async (
-  amount: string,
-  date: string,
-  description: string,
-  setAmount: React.Dispatch<React.SetStateAction<string>>,
-  setDate: React.Dispatch<React.SetStateAction<string>>,
-  setDescription: React.Dispatch<React.SetStateAction<string>>
-) => {
-  if (!description || !amount || !date) {
-    Toast.show({
-      type: "error",
-      text1: "Error adding transaction.",
-      text2: "Please try again.",
-    });
-    return;
-  }
-
-  if (!validateFormInputs(amount, date, description)) {
-    Toast.show({
-      type: "error",
-      text1: "Error adding transaction.",
-      text2: "Please try again.",
-    });
-    return;
-  }
-
-  const firebase = await getFireApp();
-  if (!firebase) {
-    throw new Error("Firebase app not initialized");
-  }
-
-  const db = firebase.firestore();
-  const uid = firebase.auth().currentUser?.uid;
-  if (!uid) {
-    throw new Error("User not authenticated");
-  }
-
-  const transactionsRef = db
-    .collection("users")
-    .doc(uid)
-    .collection("transactions");
-
-  amount = Number(amount).toFixed(2);
-
-  const numericAmount = parseFloat(amount).toFixed(2);
-  if (isNaN(numericAmount as any)) {
-    Toast.show({
-      type: "error",
-      text1: "Invalid amount.",
-      text2: "Please enter a valid number.",
-    });
-    return;
-  }
-
-  const newTransaction: Omit<TransactionProps, "id"> = {
-    date: parseDate(date),
-    description,
-    amount: numericAmount,
-  };
-
-  await transactionsRef.add(newTransaction);
-  setAmount("");
-  setDate("");
-  setDescription("");
-};
-
-/**
- * Fetches the transactions for the current user from Firestore.
- * @returns An array of transactions.
- */
-export const fetchTransactions = async (
-  setLoading: React.Dispatch<React.SetStateAction<boolean>>
-): Promise<TransactionProps[]> => {
-  setLoading(true);
-  const firebase = await getFireApp();
-  if (!firebase) {
-    throw new Error("Firebase app not initialized");
-  }
-  const db = firebase.firestore();
-  const uid = firebase.auth().currentUser?.uid;
-  if (!uid) {
-    throw new Error("User not authenticated");
-  }
-
-  const transactionsRef = db
-    .collection("users")
-    .doc(uid)
-    .collection("transactions");
-  const snapshot = await transactionsRef.orderBy("date", "desc").get();
-  let transactions: TransactionProps[] = snapshot.docs.map((doc) => ({
-    id: doc.id,
-    ...doc.data(),
-  })) as TransactionProps[];
-
-  if (!transactions) {
-    transactions = [];
-  }
-  setLoading(false);
-  return transactions;
+export const formatDate = (timestamp: any): string => {
+  const date = timestamp.toDate();
+  const month = String(date.getMonth() + 1).padStart(2, "0");
+  const day = String(date.getDate()).padStart(2, "0");
+  const year = String(date.getFullYear()).slice(2);
+  return `${month}/${day}/${year}`;
 };
 
 /**
@@ -174,11 +78,20 @@ export const isValidAmount = (amount: string): boolean => {
   return isValid;
 };
 
+
+/**
+ * Validates the form inputs to ensure they are valid.
+ * @param amount The amount string to validate.
+ * @param date The date string to validate.
+ * @param description The description string to validate.
+ * @returns `true` if the inputs are valid, otherwise `false`.
+ */
 export const validateFormInputs = (
   amount: string,
   date: string,
   description: string
 ): boolean => {
+  // #TODO: Remove console.log statements
   console.log("amount: ", amount);
   console.log("date: ", date);
   console.log("description: ", description);
@@ -194,31 +107,28 @@ export const validateFormInputs = (
 };
 
 /**
- * Parses a date string in the format MM-DD-YY to a Date object.
+ * Parses a date string in the formats MM/DD/YY, MM-DD-YY, MMDDYY to a Date object.
  * @param dateStr The date string to parse.
  * @returns A Date object representing the parsed date.
  */
-const parseDate = (dateStr: string): Date => {
+export const parseDate = (dateStr: string): Date => {
   let month, day, year;
 
   if (dateStr.includes("/")) {
-    // Format: MM/DD/YY
     [month, day, year] = dateStr.split("/").map(Number);
   } else if (dateStr.includes("-")) {
-    // Format: MM-DD-YY
     [month, day, year] = dateStr.split("-").map(Number);
   } else {
-    // Format: MMDDYY
     if (dateStr.length !== 6) {
-      return new Date(); // Invalid format, return current date
+      return new Date();
     }
     month = Number(dateStr.slice(0, 2));
     day = Number(dateStr.slice(2, 4));
     year = Number(dateStr.slice(4, 6));
   }
 
-  const parsedYear = year + 2000; // Adjust year for '20' prefix
-  const parsedDate = new Date(parsedYear, month - 1, day); // month is 0-indexed
+  const parsedYear = year + 2000;
+  const parsedDate = new Date(parsedYear, month - 1, day);
 
   return isNaN(parsedDate.getTime()) ? new Date() : parsedDate;
 };
