+++ components/grid/Column.tsx
@@ -1,0 +1,25 @@
+import { ScaledSheet, vs } from "react-native-size-matters";
+import { AppThemedView } from "../app_components/AppThemedView";
+
+interface ColumnProps {
+  children: React.ReactNode;
+  inlineStyles?: object;
+}
+
+const Column = ({ children, inlineStyles, ...props }: ColumnProps) => {
+  return(
+    <AppThemedView style={[styles.column, inlineStyles]} {...props}>{children}</AppThemedView>
+  );
+};
+
+const styles = ScaledSheet.create({
+    column: {
+        alignItems: "flex-start",
+        display: "flex",
+        flexDirection: "column",
+        justifyContent: "center",
+        width: "33%",
+      },
+});
+
+export default Column;
